+++ script.playalbum/addon.xml
@@ -1,19 +1,19 @@
 <?xml version="1.0" encoding="UTF-8" standalone="yes"?>
-<addon id="script.playalbum" name="Play Album Script" version="4.0.1" provider-name="ronie">
+<addon id="script.playalbum" name="Play Album Script" version="4.0.2" provider-name="ronie">
 	<requires>
 		<import addon="xbmc.python" version="3.0.0"/>
 	</requires>
 	<extension point="xbmc.python.library" library="default.py" />
 	<extension point="xbmc.addon.metadata">	
-		<language></language>
 		<summary lang="en_GB">Play Album Script</summary>
 		<description lang="en_GB">Script will play a full album when executed in the album info dialog.</description>
 		<platform>all</platform>
 		<forum>https://forum.kodi.tv/showthread.php?tid=104189</forum>
 		<source>https://gitlab.com/ronie/script.playalbum</source>
-		<license>GNU GENERAL PUBLIC LICENSE Version 2</license>
+		<license>GPL-2.0-only</license>
 		<assets>
 			<icon>resources/icon.png</icon>
 		</assets>
+		<news>- code refactor</news>
 	</extension>
 </addon>
